fix(workspaces-widget): surgical insert in migration, preserves user customizations
