+++ src/__tests__/workspaces-widget-migration.test.ts
@@ -33,12 +33,13 @@ describe("migrateGroupDashboardWidgets", () => {
     vi.resetAllMocks();
   });
 
-  it("rewrites the file when it exists and lacks gnar:workspaces", async () => {
+  it("inserts gnar:workspaces before gnar:columns, preserving existing content", async () => {
+    const original =
+      "# My Project\n\nCustom intro.\n\n```gnar:columns\nchildren:\n  - name: custom\n```\n";
     (invoke as ReturnType<typeof vi.fn>).mockImplementation(
       async (cmd: string) => {
         if (cmd === "file_exists") return true;
-        if (cmd === "read_file")
-          return "# My Project\n\n```gnar:columns\nchildren:\n```\n";
+        if (cmd === "read_file") return original;
         return undefined;
       },
     );
@@ -52,8 +53,19 @@ describe("migrateGroupDashboardWidgets", () => {
       (c) => c[0] === "write_file",
     );
     expect(writeCalls).toHaveLength(1);
-    expect(writeCalls[0][1].content).toContain("gnar:workspaces");
-    expect(writeCalls[0][1].content).toContain("gnar:columns");
+    const written: string = writeCalls[0][1].content;
+    expect(written).toContain("gnar:workspaces");
+    // gnar:workspaces appears before gnar:columns
+    expect(written.indexOf("gnar:workspaces")).toBeLessThan(
+      written.indexOf("gnar:columns"),
+    );
+    // custom user content survives
+    expect(written).toContain("Custom intro.");
+    expect(written).toContain("- name: custom");
+    // path is correct
+    expect(writeCalls[0][1].path).toBe(
+      "/tmp/my-project/.gnar-term/project-dashboard.md",
+    );
   });
 
   it("skips the rewrite when the file already contains gnar:workspaces", async () => {
